move mock object creator into existing fixture file
