+++ protocol-designer/src/step-generation/test-with-flow/fixtures/robotStateFixtures.js
@@ -1,4 +1,5 @@
 // @flow
+import cloneDeep from 'lodash/cloneDeep'
 import mapValues from 'lodash/mapValues'
 import { getLabwareDefURI } from '@opentrons/shared-data'
 import {
@@ -273,4 +274,22 @@ export const getStateAndContextTempMagModules = ({
     },
   }
   return { invariantContext, robotState }
+}
+
+export const robotWithStatusAndTemp = (
+  robotState: RobotState,
+  temperatureModuleId: string,
+  status:
+    | typeof TEMPERATURE_AT_TARGET
+    | typeof TEMPERATURE_APPROACHING_TARGET
+    | typeof TEMPERATURE_DEACTIVATED,
+  targetTemperature: number | null
+): RobotState => {
+  const robot = cloneDeep(robotState)
+  robot.modules[temperatureModuleId].moduleState = {
+    type: TEMPDECK,
+    targetTemperature,
+    status,
+  }
+  return robot
 }
